조회 등록 수정 삭제 완료

+++ dieat/src/main/java/com/samsung/dieat/diet_post/command/domain/repository/DietPostRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DietPostRepository extends JpaRepository<DietPost, Long> {
+public interface DietPostRepository extends JpaRepository<DietPost, Integer> {
 }
